Visual Studio Builds: Move 2010 Projects to 2012
Since the builds of the *mm stack are going to use C++11, this means that
we can't build libsigc++ with MSVC 2005~2010 anymore, so we need to update
the Visual Studio project files to version 2012 (11.0), so that we can
still support building with Visual Studio, as it does support enough of
C++11 features that is needed.

Also clean up a little bit by dropping some of the unneeded items from the
projects and solution files.

These project files are clearly going to need more cleanup, and we should
drop the MSVC_Net2005 and MSVC_Net2008 folders, and rename MSVC_Net2010 to
MSVC_Net2012, but this commit will ensure that libsigc++ is indeed still
buildable with Visual Studio, and we can support Visual Studio 2013 and
2015 in a more efficient manner.