exception_catch.h: Generate deduce_result_type for the void specializations too.
I guess this is correct and test_exception_catch still passes.
Without this, it's actually using, for instance:
  sigc::deduce_result_type<T_arg1, T_Arg2>
even though sigc::deduce_result_type<> should really be called with a result type
first, like so:
  sigc::deduce_result_type<T_return, T_arg1, T_Arg2>

Explicitly changing it to use
  sigc::deduce_result_type<T_return, T_Arg1, T_Arg2>
or
  sigc::deduce_result_type<T_Arg1, T_Arg2>
instead of just
  deduce_result_type<T_Arg1, T_Arg2>
doesn't seem to break things either, probably out of luck,
but changing it to define a type that it actually seems to expect seems
cleaner, and makes it easier for us to change sigc::deduce_result_type.