MSVC Projects: Overhaul The Projects
The former MSVC 2010 (2012) projects carried over quite a bit of cruft from
the Visual Studio 2003 days, and there are a number of items that could be
consolidated into ptoperty sheets, so it's time that the projects get a
clean up and overhaul.  This is not unlike what was done to the glibmm
projects some time ago.

Consolidate the projects all into MSVC_Net2010, without subdirectories, as
the property sheets will separate the object files of the various projects
into their own places.

This will also likely enable us to support MSVC 2015 in a more efficient
way.

Also, move the projects configs to MSVC 2013, as that is now the minimum
version of MSVC required to build sigc++.