MSVC Builds: Add "Install" Project
This makes building the -mm stack easier by copying the headers and the
built DLL, PDB and LIB files to a location where it can be picked up by
the gtkmm stack automatically, when they are extracted in a common
directory.  This is what is currently done with the GTK+ and Clutter
stacks for Visual Studio builds.