Use mem_fun.h as a regular .h file.
Instead of generating it from an .m4 file.
This still has some repetition, but it is not nearly as much as before,
and it is the last m4 file.
Having all the code in normal C++ source files makes it far easier for people
to contribute, to actually reduce this repetition.