Build: Use NMake Makefiles for MSVC Builds
Since the MSVC projects are not created upon 'make dist' but
is manually updated, use a set of NMake Makefiles that consumes
sigc++/filelist.am as needed, so that whenever the list of sources
is updated, we do not have to worry that the generated release
tarballs have broken Visual Studio build files.

The next commit will remove the Visual Studio project files.

This should make maintenance of the Visual Studio build files easier.