builds: Add NMake Makefiles
This adds a set of NMake Makefiles that can be used to build
libsigc++-2.10.x (and the later C++11 versions of libsigc++) with Visual
Studio 2013 or later.  Building the example, the tests and the
benchmarking programs are supported in addition to building the main
libsigc++ DLL.

Note that for the C++-11 releases, we name the DLLs and LIBs as
sigc-vc140-2_0.[dll|lib] or sigc-vc140-d-2_0.[dll|lib] for Visual Studio
2015 and 2017 builds as these builds link to the msvcp140[d].dll and
vcruntime140[d].dll C/C++ runtime DLLs.

This set of NMake Makefiles are now dist'ed in place of the Visual
Studio 2013 project files.