meson/Windows: Fix builds when builddir is a subdir of source tree
It appears that Meson did not construct the paths properly for
shutil.copy2() when building in a build directory that is a
subdirectory of the sources, when using meson.current_build_dir(),
possibly due how path separators are handled.

Fix this by constructing the paths using project_build_root / 'MSVC_NMake'.