meson.build: Fix 'meson dist' on Windows
Since we are assured that we are using Python 3.x on when we run Meson, we
do not really need to look for the 'python3' executable, but we could
just use whatever Python interpreter that is used to run Meson.

This will fix situations where it is commonly the case where we may have
multiple Python 3.x installations on Windows (www.python.org, and those
from Cygwin/mingw-w64), so that Meson really uses one and only one
Python installation to run everything that is Python-related, which will
thus fix '[meson|ninja] dist' on Visual Studio builds.