NMake Makefiles: Use toolset version in DLL filename by default
The libsigc++ DLL and .lib was only named 'sigc-vc150(d)-3_0' until the
previous patches in this series improved Visual Studio 2019 support to
name the DLL and .lib by using 'sigc-vc160(d)-3_0'.

This updates the naming of the DLL and .lib to the toolset version as
defined by Microsoft by default, i.e. 'vc141' for Visual Studio 2017
and 'vc142' for Visual Studio 2019.  If the former DLL naming ('vc150')
is desired, passing 'USE_COMPAT_LIBS=1' will achive this.