NMake Makefiles: Add rules to generate sigc.rc and sigc++config.h
This adds a 'prep-git-build' target which is run before building any of the
other targets so that this paves the first steps in making the libsigc++-2-10
branch buildable directly from a GIT checkout with the NMake Makefiles.

Note that unlike the master (3.x) branch, we still need to use m4 to generate
some sources and headers, which will be covered in the following commit.