Meson build: Fix dependency on generated files
Add meson.build files in all subdirectories of sigc++/ where .h and/or .cc
files are generated. Looks like it's necessary in order to have all
dependent .cc files recompiled when a .h.m4 file has been changed.
Don't know if it has always been necessary, or if it has become necessary
due to changes in Meson and/or Ninja.