build: Actually support MSVC static builds
Apply -DLIBSIGCXX_STATIC as appropriate when we request a static build to be
done for the Meson and NMake builds, and skip building the version .rc file
if a static build is requested.

For the NMake builds, separate the build artifacts from the static and shared
builds.

The CMake builds are not updated here as it always assumes a shared build, nor
are the autotools builds since it is not used for Visual Studio builds at all.